+++ src/Elems/MainElems/Candidate.tsx
@@ -5,6 +5,7 @@ import { IndexToNine, _ReactProps } from '../../Types';
 
 type CandidateProps = Readonly<{
    index: IndexToNine
+   className?: string
 }> & _ReactProps
 
 /**
@@ -29,6 +30,10 @@ export default class Candidate extends React.Component<CandidateProps> {
       let thisClassName = "Candidate"
       if (typeof this.props.children === "string" || typeof this.props.children === "number") {
          thisClassName += ` digit-${Number(this.props.index) + 1}`
+      }
+
+      if (this.props.className) {
+         thisClassName += ` ${this.props.className}`
       }
 
       return (